Split compositing/pipeline/constellation code into separate crate.
