+++ configure
@@ -604,6 +604,7 @@ done
 
 make_dir ${CFG_BUILD_DIR}src/components/macros
 make_dir ${CFG_BUILD_DIR}src/components/util
+make_dir ${CFG_BUILD_DIR}src/components/compositing
 make_dir ${CFG_BUILD_DIR}src/components/embedding
 make_dir ${CFG_BUILD_DIR}src/components/msg
 make_dir ${CFG_BUILD_DIR}src/components/net
