+++ src/components/compositing/pipeline.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use compositing::CompositorChan;
+use CompositorChan;
 use layout::layout_task::LayoutTask;
 
 use gfx::render_task::{PaintPermissionGranted, PaintPermissionRevoked};
