+++ src/components/compositing/quadtree.rs
