+++ src/components/main/servo.rs
@@ -13,50 +13,20 @@ extern crate log;
 
 extern crate debug;
 
-extern crate alert;
-extern crate azure;
-extern crate geom;
-extern crate gfx;
-#[cfg(not(target_os="android"))]
-extern crate glfw;
-#[cfg(target_os="android")]
-extern crate glut;
-extern crate js;
-extern crate layers;
-extern crate layout;
-extern crate opengles;
-extern crate png;
+extern crate compositing;
 extern crate rustuv;
-extern crate script;
-#[phase(plugin)]
-extern crate servo_macros = "macros";
 extern crate servo_net = "net";
 extern crate servo_msg = "msg";
 #[phase(plugin, link)]
 extern crate servo_util = "util";
-extern crate style;
-extern crate sharegl;
-extern crate stb_image;
-
-extern crate collections;
 extern crate green;
 extern crate libc;
 extern crate native;
 extern crate rustrt;
-extern crate serialize;
-extern crate sync;
-extern crate time;
 extern crate url;
 
-#[cfg(target_os="macos")]
-extern crate core_graphics;
-#[cfg(target_os="macos")]
-extern crate core_text;
-
 #[cfg(not(test))]
-use compositing::{CompositorChan, CompositorTask};
-#[cfg(not(test))]
-use constellation::Constellation;
+use compositing::{CompositorChan, CompositorTask, Constellation};
 #[cfg(not(test))]
 use servo_msg::constellation_msg::{ConstellationChan, InitLoadUrlMsg};
 
@@ -84,17 +54,6 @@ use rustrt::task::TaskOpts;
 #[cfg(not(test))]
 use url::Url;
 
-
-#[path="compositing/compositor_task.rs"]
-pub mod compositing;
-
-pub mod constellation;
-pub mod pipeline;
-
-pub mod windowing;
-
-#[path="platform/mod.rs"]
-pub mod platform;
 
 #[cfg(not(test), target_os="linux")]
 #[cfg(not(test), target_os="macos")]
